Factored out graph.processor.emf
Mapping of graph elements to Ecore using properties is fragile.
Currently there is a stack overflow error in setting parent, whcih is
not worth solving.

Moved org.nasdanika.graph.processor.emf to a new module which is not
part of the parent pom module
list.

The module can be added to the parent pom when needed.